requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
pydantic==2.9.0
pydantic-settings==2.3.4
langchain==0.2.11
langchain-community==0.2.10
langchain-google-genai==1.0.6
google-generativeai==0.7.2
chromadb==0.5.3
sentence-transformers==3.0.1
pypdf==4.2.0
ollama==0.3.1
pyttsx3==2.90
openai-whisper==20231117
streamlit==1.36.0
requests==2.32.3
python-dotenv==1.0.1
pytest==8.3.2
httpx==0.27.0
